modify one_relu model to use int16

+++ tvm_practice/models/real_model2.py
@@ -94,7 +94,7 @@ def getOneConvModel():
 
 def getOneReluModel():
   # input_ = relay.var("input", shape=(1, 28, 4, 4))
-  input_ = relay.var("input", shape=(1,28,4,4), dtype="int8")
+  input_ = relay.var("input", shape=(1,28,4,4), dtype="int16")
   y = relay.nn.pad(input_, pad_width=((0, 0),(0, 0),(1, 1),(1,1)), pad_value=0)
   y = relay.nn.relu(y)
 
@@ -103,7 +103,7 @@ def getOneReluModel():
     # "quant_zp": np.random.randint(0, 255, 28).astype("int"),
     # "weight": np.random.rand(28,28,3,3).astype("float32")
   }
-  
+
 
   out = tvm.IRModule.from_expr(y)
